refactor(run-protocol): move Collect out of bootstrapRunLoc.js
