+++ packages/run-protocol/src/bootstrapRunLoC.js
@@ -1,36 +1,8 @@
 // @ts-check
 import { E } from '@agoric/eventual-send';
-import { makeGovernedRatio, ParamType } from '@agoric/governance';
+import { makeGovernedRatio } from '@agoric/governance';
 import { CreditTerms } from './getRUN.js';
 import { makeElectorateParams } from './vaultFactory/params.js';
-
-const { entries, fromEntries, keys, values } = Object;
-
-export const Collect = {
-  /**
-   * @param {Record<string, V>} obj
-   * @param {(v: V) => U} f
-   * @returns {Record<string, U>}
-   * @template V
-   * @template U
-   */
-  mapValues: (obj, f) => fromEntries(entries(obj).map(([p, v]) => [p, f(v)])),
-  /**
-   * @param {X[]} xs
-   * @param {Y[]} ys
-   * @returns {[X, Y][]}
-   * @template X
-   * @template Y
-   */
-  zip: (xs, ys) => xs.map((x, i) => [x, ys[i]]),
-  /**
-   * @param {Record<string, ERef<V>>} obj
-   * @returns {Promise<Record<string, V>>}
-   * @template V
-   */
-  allValues: async obj =>
-    fromEntries(Collect.zip(keys(obj), await Promise.all(values(obj)))),
-};
 
 /**
  * @typedef { import('@agoric/eventual-send').Unpromise<T> } Unpromise<T>
